fixing some warnings
git-svn-id: svn://192.168.1.10/public/CoffeeMud@16947 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Items/Software/RocketShipProgram.java
@@ -584,7 +584,7 @@ public class RocketShipProgram extends GenShipProgram
 		{
 			for(final ShipEngine engineE : programEngines)
 				performSimpleThrust(engineE,newInject, true);
-			if((CMath.abs(targetAccelleration)-this.lastAccelleration)<.01)
+			if((CMath.abs(targetAccelleration)-this.lastAccelleration.doubleValue())<.01)
 				break;
 			newInject = this.calculateMarginalTargetInjection(this.lastInject, targetAccelleration);
 		}
